TS compiles and it loads correctly in browser
