--- client/dash.js
@@ -1,2 +1,0 @@
-"use strict";
-document.body.appendChild(document.createElement('div'));
