+++ client/tsconfig.json
@@ -1,16 +1,24 @@
 {
   "compilerOptions": {
     "target": "ES2015",
-    "module": "commonjs",
+    "module": "ES6",
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "pretty": true,
     "skipLibCheck": true,
+    "rootDirs": [
+      "./",
+      "node_modules",
+      "node_modules/@types"
+    ],
+    "rootDir": "src",
+    "outDir": "./",
     "lib": [
       "es2017",
       "esnext.asynciterable",
-      "DOM"
+      "dom"
     ]
-  }
+  },
+  "exclude": []
 }
